Move FakeEventStream to correct package
